Add files via upload
Almost done with this calculator app. I learned a lot of new tricks from this for HTML and CSS as well as JS. I first ran into an issue where I forgot to add a function so I kept getting error messages while trying to use certain buttons on the calculator. The next error I made in following along was not noticing they had switched tabs and were showing a current project view and finished program view on 2 separate tabs depending on what they were explaining. I thought my calculator was broken because it was not behaving the same as the instructor's but they were showing how it SHOULD look.